Changes interface for graph
Changed how edges are done in the multigraph to use a class that is
passed in. Interface has changed so uses of it need changed as well.

Added interface for directed edge to define functionality of edges
that go into the graph.

Added Track as concrete implementation of directededge for the metro
system.

Implementation of generategraphfromfile complete in the parser.

Tests need added and updated.